fix(training): campagna ML certificata — no leak, storico pieno, 5 shard, robusta
Workflow (retrain_models.yml): i run da cron usavano fallback errati
(3 stagioni, budget 150, 3 shard). Corretti i fallback: LAST_N_SEASONS=20
(storico pieno), TIME_BUDGET_MIN=240 (4h), TOTAL_SHARDS=5 (5 leghe in parallelo).

Leak standings (CRITICO): la tabella standings e' uno snapshot di FINE stagione;
veniva unita a OGNI partita esponendo home_rank/points/win/lose/played/goals_*
(18 colonne) = leak del risultato finale (anche di partite future). Il drop_cols
ne filtrava solo le 3 colonne-chiave. Ora tutte le colonne standings sono
prefissate standings_ => eliminate davvero dalle feature.

Target completi: corner/tiri-in-porta/cartellini/timing-gol erano degeneri in
pre_match (label lette da colonne assenti). Ora le statistiche/eventi della
partita stessa vengono unite per costruire le label (riusando le righe gia'
scaricate, zero I/O extra) e rimosse dalle feature via drop_cols (solo y, no leak).
Abilita 2 mercati binari prima persi (first_goal_before_30, goal_in_2h).

Fix ht_over_0_5: (bool).where() produceva dtype object (bool+NaN) => crash
"Unknown label type" su leghe con dati primo-tempo parziali. Cast a float64.

Allineamento train/predict: predict ora esclude i giocatori (parita' col training)
e calcola ELO/H2H su storico pieno (matches leggeri) come il training, evitando
lo skew dell'ELO cumulativo; eventi/statistiche restano a finestra corta in serving.

Robustezza: guardie try/except sui passi di arricchimento opzionali (label, form)
oltre ai 3 livelli esistenti (shard/lega/target) => nessuna lega fa crashare lo
shard. Testato su lega minima (50 partite) e piena (600, 10 stagioni): 0 crash.

Planner: cutoff a 2026-06-16T08:00Z => tutte le 1014 leghe eleggibili (>=50
partite giocate) vengono riaddestrate con la metodologia corretta; convergenza a 0.

+++ Ai Engine/ai_engine/targets.py
@@ -79,8 +79,13 @@ def add_targets_from_matches(df: pd.DataFrame) -> pd.DataFrame:
     # First Half Over 0.5 (direct boolean target)
     # Use .where() to propagate NaN when halftime data is missing, instead of
     # silently coercing NaN > 0 to False (which creates false negatives).
+    # FIX 2026-06-16: cast a FLOAT (1.0/0.0/NaN) PRIMA di .where(). Un
+    # `(serie_bool).where(mask)` produce dtype `object` (bool + NaN) e quando la
+    # lega ha dati primo-tempo PARZIALI il classificatore lo rifiuta con
+    # "Unknown label type: unknown" => il mercato prioritario ht_over_0_5 non
+    # veniva addestrato. Con float, dopo dropna restano 1.0/0.0 => binary pulito.
     ht_total = out["halftime_home"] + out["halftime_away"]
-    out["target_ht_over_0_5"] = (ht_total > 0).where(ht_total.notna())
+    out["target_ht_over_0_5"] = (ht_total > 0).astype("float64").where(ht_total.notna())
 
     # Exact score — keep NaN when goals are missing instead of using "-1--1"
     has_goals = out["goals_home"].notna() & out["goals_away"].notna()
